Update deploy and execute test
